requirements.txt
discord.py
requests
pystyle